requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
Appium-Python-Client==1.3.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
concurrencytest==0.1.2
configparser==5.2.0
cryptography==36.0.0
extras==1.0.0
fixtures==3.0.0
frozenlist==1.3.0
h11==0.12.0
idna==3.3
multidict==6.0.2
outcome==1.1.0
pbr==5.8.0
pycparser==2.21
pyOpenSSL==21.0.0
python-subunit==1.4.0
requests==2.26.0
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
testtools==2.5.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0
yarl==1.7.2